Add Music and Music tests apps to ATF test packages
Since the packages are not part of the ATF repository, they should be
added to the ATF package list through their own makefiles.

This patch is necessary to allow an out of the box inclusion of the
ATF repository in any tree (like the AOSP upstream/master tree).

Issue: AXIA-2227
Change-Id: Ib6970a725b64a1aa4200758b6ac237aaf1d6b16e
Signed-off-by: Ceausu Mihai <mihai.ceausu@intel.com>